requirements.txt
# Core dependencies
requests>=2.32.0
pandas>=2.3.0
mplfinance>=0.12.10b0
tqdm>=4.66.0
python-dotenv>=1.0.0

# Web framework
Flask>=3.0.0
Flask-SocketIO>=5.3.0

# Background tasks
celery>=5.3.0
redis>=5.0.0

# Additional utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Plotting and visualization
plotly==6.3.0
kaleido==1.1.0
narwhals==2.5.0
